+++ datasette_dashboards/static/dashboards.js
@@ -201,3 +201,8 @@ function toggleFullscreen() {
     el.requestFullscreen()
   }
 }
+
+function autorefresh(minutes) {
+  const timeout = Math.round(minutes * 60 * 1000)
+  window.setTimeout(function () { window.location.reload() }, timeout)
+}